+++ tests/test_coroutine_listeners.py
@@ -22,9 +22,23 @@ def add_async_mock_listener(ee):
 
 
 @pytest.mark.asyncio
-async def test_coroutine_listener_called(ee, add_async_mock_listener):
+async def test_coroutine_listener_called(
+    ee, add_async_mock_listener, event_loop
+):
+    ee.set_event_loop(event_loop)
     async_foo_listener = add_async_mock_listener("foo")
     ee.emit("foo")
     # Just to allow other queued coroutines to run
     await asyncio.sleep(0)
+    async_foo_listener.assert_awaited_once_with()
+
+
+@pytest.mark.asyncio
+async def test_emitting_from_different_thread(
+    ee, add_async_mock_listener, event_loop
+):
+    ee.set_event_loop(event_loop)
+    async_foo_listener = add_async_mock_listener("foo")
+
+    await asyncio.to_thread(ee.emit, "foo")
     async_foo_listener.assert_awaited_once_with()